Preserve the preview's aspect ratio.
Still doesn't work in portrait right... The way TileBrush does its
RelativeTransform is bizarre. A rotation that changes the brush's
dimensions (i.e. 90 degree rotation of a rectangle) causes it to be
projected onto the target canvas in a distorted way...

See http://msdn.microsoft.com/en-us/library/aa970500.aspx

The scale for that composite transform will need to be different when
it's in portrait to compensate for this fact, or I'll have to abandon
using a relative transform, which means the sizes will need to be
calculated, etc. etc., which will be a Fucking Headache.